add temporary conversations and use them for group chats
Add temporary conversations that are not (yet) created by the backends
and, thus, do not time out and get removed by the backends. When a
backend adds a conversation that matches a temporary conversation, the
temporary conversation is converted into a regular conversation.

When the user joins a group chat, create a temporary conversation for
the chat. When the user parts a temporary group chat conversation, close
the conversation immediately.

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>