do not log messages to uninitialized conversation log
Do not log message to a conversation with the log() method, if the
conversation has not been initialized yet. Messages must be loaded from
history file first to avoid showing messages that arrived before loading
the history file twice in the log (at the beginning and at the end of
the log).

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>