add conversation list filtering
Add filtering to the conversation list and use the GO_CONV keybind for
it. By default, this is bound to the CTRL-V keys.

Filtering allows the specification of a filter string and selecting
matches with the UP and DOWN keys. Non-matching conversations are shown
as dim text. Filter mode can be left with the ESC key or by entering a
conversation.

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>