treat files in PATH starting with "nuqql-" as backends
Treat files in PATH starting with "nuqql-" as backends if they are not
blacklisted in BACKEND_BLACKLIST in backend.py. Remove the
start_based(), start_matrixd() and start_slixmppd() functions because we
do not need them anymore. Purpled is still handled the old way because
it uses different command line arguments.

Signed-off-by: hwipl <33433250+hwipl@users.noreply.github.com>